# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koperasi App API",
        "description": "RESTful API for Koperasi App mobile applications and integrations",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dashboard/stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get member dashboard statistics",
                "description": "Get dashboard statistics",
                "operationId": "cabcc6ba02fd4175bde81636208c73d8",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "anggota": {
                                            "properties": {
                                                "nama": {
                                                    "type": "string",
                                                    "example": "Budi Santoso"
                                                },
                                                "no_anggota": {
                                                    "type": "string",
                                                    "example": "ANG-2026-0001"
                                                },
                                                "lama_keanggotaan_label": {
                                                    "type": "string",
                                                    "example": "6 bulan"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "totalSimpanan": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 320000
                                        },
                                        "simpananPokok": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 50000
                                        },
                                        "simpananWajib": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 270000
                                        },
                                        "limitMaksimal": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 960000
                                        },
                                        "limitTersedia": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 960000
                                        },
                                        "sisaAngsuranAktif": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "cicilanPokokAktif": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 0
                                        },
                                        "pinjamanAktif": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nominal": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "tenor_bulan": {
                                                    "type": "integer"
                                                },
                                                "sisa_angsuran": {
                                                    "type": "integer"
                                                },
                                                "total_angsuran": {
                                                    "type": "integer"
                                                },
                                                "sisa_total_bayar": {
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Anggota not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/actionable": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get actionable items (for admin dashboard)",
                "description": "Get actionable items",
                "operationId": "a6edecce88435cc65127c542d96a9ceb",
                "responses": {
                    "200": {
                        "description": "Actionable items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/charts": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get 6-month trend chart data",
                "description": "Get chart data",
                "operationId": "acf5b2ff4cfbada8aa78149ff892ed77",
                "responses": {
                    "200": {
                        "description": "Chart data for trends",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "labelBulan": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "grafikTren": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "bulan": {
                                                        "type": "string",
                                                        "example": "Mar 2026"
                                                    },
                                                    "simpanan": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "pinjaman": {
                                                        "type": "number",
                                                        "format": "float"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Anggota not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/aktivitas": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get recent activity (loan applications & payments)",
                "description": "Get recent activity",
                "operationId": "be4f4eb439579d2cb658a16f7e23067c",
                "responses": {
                    "200": {
                        "description": "Recent activity list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "tipe": {
                                                "type": "string",
                                                "enum": [
                                                    "pinjaman",
                                                    "angsuran"
                                                ]
                                            },
                                            "nama": {
                                                "type": "string"
                                            },
                                            "keterangan": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "tanggal_format": {
                                                "type": "string",
                                                "example": "23 Aug 2026"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Anggota not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}